+++ app/src/main/java/com/gomson/tryangle/domain/LineGuide.java
@@ -2,7 +2,7 @@ package com.gomson.tryangle.domain;
 
 public class LineGuide extends Guide {
 
-    public LineGuide(int objectId, int guideId) {
+    public LineGuide(long objectId, int guideId) {
         super(objectId, guideId);
     }
 }
