+++ app/src/main/java/com/gomson/tryangle/domain/ObjectGuide.java
@@ -6,7 +6,7 @@ public class ObjectGuide extends Guide {
     private int diffY;
     private int objectClass;
 
-    public ObjectGuide(int objectId, int guideId, int diffX, int diffY, int objectClass) {
+    public ObjectGuide(long objectId, int guideId, int diffX, int diffY, int objectClass) {
         super(objectId, guideId);
         this.diffX = diffX;
         this.diffY = diffY;
